Python3.6
This is fleet information querying system for logistic companies used for generating insights and
perform analytics of their fleet operations by the managers and executives of the company. The
users can easily query the application to find the status of the fleets and analytics of fleet
operations with English like queries.